+++ naga/src/back/msl/mod.rs
@@ -838,6 +838,7 @@ pub fn supported_capabilities() -> crate::valid::Capabilities {
         // No DRAW_INDEX
         // No MEMORY_DECORATION_VOLATILE
         | Caps::MEMORY_DECORATION_COHERENT
+        | Caps::DEBUG_PRINTF
 }
 
 #[test]