+++ naga/src/back/msl/writer.rs
@@ -4273,6 +4273,21 @@ impl<W: Write> Writer<W> {
                     writeln!(self.out, ");")?;
                 }
                 crate::Statement::RayPipelineFunction(_) => unreachable!(),
+                crate::Statement::DebugPrintf {
+                    ref format,
+                    ref arguments,
+                } => {
+                    write!(
+                        self.out,
+                        "{level}metal::os_log_default.log_info(\"{}\"",
+                        format
+                    )?;
+                    for &arg in arguments {
+                        write!(self.out, ", ")?;
+                        self.put_expression(arg, &context.expression, true)?;
+                    }
+                    writeln!(self.out, ");")?;
+                }
             }
         }