+++ naga/src/back/pipeline_constants.rs
@@ -921,6 +921,14 @@ fn adjust_stmt(new_pos: &HandleVec<Expression, Handle<Expression>>, stmt: &mut S
                 adjust(payload);
             }
         },
+        Statement::DebugPrintf {
+            format: _,
+            ref mut arguments,
+        } => {
+            for argument in arguments.iter_mut() {
+                adjust(argument);
+            }
+        }
         Statement::Break
         | Statement::Continue
         | Statement::Kill