+++ naga/src/front/spv/mod.rs
@@ -1664,6 +1664,7 @@ impl<I: Iterator<Item = u32>> Frontend<I> {
                 }
                 S::WorkGroupUniformLoad { .. } => unreachable!(),
                 S::CooperativeStore { .. } => unreachable!(),
+                S::DebugPrintf { .. } => unreachable!(),
             }
             i += 1;
         }