+++ naga/src/ir/mod.rs
@@ -2361,6 +2361,11 @@ pub enum Statement {
         target: Handle<Expression>,
         data: CooperativeData,
     },
+    /// This corresponds to `debugPrintf` in WGSL when the `wgpu_debug_printf` extension is enabled.
+    DebugPrintf {
+        format: String,
+        arguments: Vec<Handle<Expression>>,
+    },
 }
 
 /// A function argument.