+++ wgpu-naga-bridge/src/lib.rs
@@ -179,6 +179,11 @@ pub fn features_to_naga_capabilities(
         Caps::MEMORY_DECORATION_VOLATILE,
         features.contains(wgt::Features::MEMORY_DECORATION_VOLATILE),
     );
+
+    caps.set(
+        Caps::DEBUG_PRINTF,
+        features.contains(wgt::Features::DEBUG_PRINTF),
+    );
     caps
 }